SaveMapCommand scale must be greater than 0. (#3724)

+++ src/main/java/meteordevelopment/meteorclient/commands/commands/SaveMapCommand.java
@@ -63,7 +63,7 @@ public class SaveMapCommand extends Command {
             saveMap(map, state, path, 128);
 
             return SINGLE_SUCCESS;
-        }).then(argument("scale", IntegerArgumentType.integer()).executes(context -> {
+        }).then(argument("scale", IntegerArgumentType.integer(1)).executes(context -> {
             int scale = IntegerArgumentType.getInteger(context, "scale");
 
             MapState state = getMapState();
